+++ spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/web/ServerProperties.java
@@ -23,60 +23,17 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.SessionCookieConfig;
 import javax.servlet.SessionTrackingMode;
 
-import io.undertow.Undertow.Builder;
-import io.undertow.UndertowOptions;
-import org.apache.catalina.Context;
-import org.apache.catalina.connector.Connector;
-import org.apache.catalina.valves.AccessLogValve;
-import org.apache.catalina.valves.RemoteIpValve;
-import org.apache.coyote.AbstractProtocol;
-import org.apache.coyote.ProtocolHandler;
-import org.apache.coyote.http11.AbstractHttp11Protocol;
-import org.eclipse.jetty.server.AbstractConnector;
-import org.eclipse.jetty.server.ConnectionFactory;
-import org.eclipse.jetty.server.Handler;
-import org.eclipse.jetty.server.HttpConfiguration;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.server.handler.ContextHandler;
-import org.eclipse.jetty.server.handler.HandlerCollection;
-import org.eclipse.jetty.server.handler.HandlerWrapper;
-
-import org.springframework.boot.autoconfigure.web.ServerProperties.Session.Cookie;
-import org.springframework.boot.cloud.CloudPlatform;
 import org.springframework.boot.context.embedded.Compression;
-import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizerBeanPostProcessor;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerFactory;
-import org.springframework.boot.context.embedded.InitParameterConfiguringServletContextInitializer;
 import org.springframework.boot.context.embedded.Servlet;
 import org.springframework.boot.context.embedded.Ssl;
-import org.springframework.boot.context.embedded.jetty.JettyEmbeddedServletContainerFactory;
-import org.springframework.boot.context.embedded.jetty.JettyServerCustomizer;
-import org.springframework.boot.context.embedded.tomcat.TomcatConnectorCustomizer;
-import org.springframework.boot.context.embedded.tomcat.TomcatContextCustomizer;
-import org.springframework.boot.context.embedded.tomcat.TomcatEmbeddedServletContainerFactory;
-import org.springframework.boot.context.embedded.undertow.UndertowBuilderCustomizer;
-import org.springframework.boot.context.embedded.undertow.UndertowEmbeddedServletContainerFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.boot.context.properties.DeprecatedConfigurationProperty;
 import org.springframework.boot.context.properties.NestedConfigurationProperty;
-import org.springframework.boot.web.servlet.ServletContextInitializer;
-import org.springframework.context.EnvironmentAware;
-import org.springframework.core.Ordered;
-import org.springframework.core.env.Environment;
-import org.springframework.util.ObjectUtils;
-import org.springframework.util.StringUtils;
 
 /**
- * {@link ConfigurationProperties} for a web server (e.g. port and path settings). Will be
- * used to customize an {@link EmbeddedServletContainerFactory} when an
- * {@link EmbeddedServletContainerCustomizerBeanPostProcessor} is active.
+ * {@link ConfigurationProperties} for a web server (e.g. port and path settings).
  *
  * @author Dave Syer
  * @author Stephane Nicoll
@@ -87,10 +44,10 @@ import org.springframework.util.StringUtils;
  * @author Quinten De Swaef
  * @author Venil Noronha
  * @author Aurélien Leboulanger
+ * @author Brian Clozel
  */
 @ConfigurationProperties(prefix = "server", ignoreUnknownFields = true)
-public class ServerProperties
-		implements EmbeddedServletContainerCustomizer, EnvironmentAware, Ordered {
+public class ServerProperties {
 
 	/**
 	 * Server HTTP port.
@@ -149,69 +106,6 @@ public class ServerProperties
 
 	private final Undertow undertow = new Undertow();
 
-	private Environment environment;
-
-	@Override
-	public int getOrder() {
-		return 0;
-	}
-
-	@Override
-	public void setEnvironment(Environment environment) {
-		this.environment = environment;
-	}
-
-	@Override
-	public void customize(ConfigurableEmbeddedServletContainer container) {
-		if (getPort() != null) {
-			container.setPort(getPort());
-		}
-		if (getAddress() != null) {
-			container.setAddress(getAddress());
-		}
-		if (getServlet().getContextPath() != null) {
-			container.setContextPath(getServlet().getContextPath());
-		}
-		if (getDisplayName() != null) {
-			container.setDisplayName(getDisplayName());
-		}
-		if (getSession().getTimeout() != null) {
-			container.setSessionTimeout(getSession().getTimeout());
-		}
-		container.setPersistSession(getSession().isPersistent());
-		container.setSessionStoreDir(getSession().getStoreDir());
-		if (getSsl() != null) {
-			container.setSsl(getSsl());
-		}
-		if (getServlet().getJsp() != null) {
-			container.setJsp(getServlet().getJsp());
-		}
-		if (getCompression() != null) {
-			container.setCompression(getCompression());
-		}
-		container.setServerHeader(getServerHeader());
-		if (container instanceof TomcatEmbeddedServletContainerFactory) {
-			getTomcat().customizeTomcat(this,
-					(TomcatEmbeddedServletContainerFactory) container);
-		}
-		if (container instanceof JettyEmbeddedServletContainerFactory) {
-			getJetty().customizeJetty(this,
-					(JettyEmbeddedServletContainerFactory) container);
-		}
-
-		if (container instanceof UndertowEmbeddedServletContainerFactory) {
-			getUndertow().customizeUndertow(this,
-					(UndertowEmbeddedServletContainerFactory) container);
-		}
-		container.addInitializers(new SessionConfiguringInitializer(this.session));
-		container.addInitializers(new InitParameterConfiguringServletContextInitializer(
-				getServlet().getContextParameters()));
-	}
-
-	public void setLoader(String value) {
-		// no op to support Tomcat running as a traditional container (not embedded)
-	}
-
 	public Integer getPort() {
 		return this.port;
 	}
@@ -258,14 +152,6 @@ public class ServerProperties
 
 	public void setMaxHttpHeaderSize(int maxHttpHeaderSize) {
 		this.maxHttpHeaderSize = maxHttpHeaderSize;
-	}
-
-	protected final boolean getOrDeduceUseForwardHeaders() {
-		if (this.useForwardHeaders != null) {
-			return this.useForwardHeaders;
-		}
-		CloudPlatform platform = CloudPlatform.getActive(this.environment);
-		return (platform == null ? false : platform.isUsingForwardHeaders());
 	}
 
 	public Integer getConnectionTimeout() {
@@ -684,6 +570,14 @@ public class ServerProperties
 			this.maxConnections = maxConnections;
 		}
 
+		public int getMaxHttpHeaderSize() {
+			return this.maxHttpHeaderSize;
+		}
+
+		public void setMaxHttpHeaderSize(int maxHttpHeaderSize) {
+			this.maxHttpHeaderSize = maxHttpHeaderSize;
+		}
+
 		public int getAcceptCount() {
 			return this.acceptCount;
 		}
@@ -698,212 +592,6 @@ public class ServerProperties
 
 		public void setAdditionalTldSkipPatterns(List<String> additionalTldSkipPatterns) {
 			this.additionalTldSkipPatterns = additionalTldSkipPatterns;
-		}
-
-		void customizeTomcat(ServerProperties serverProperties,
-				TomcatEmbeddedServletContainerFactory factory) {
-			if (getBasedir() != null) {
-				factory.setBaseDirectory(getBasedir());
-			}
-			factory.setBackgroundProcessorDelay(Tomcat.this.backgroundProcessorDelay);
-			customizeRemoteIpValve(serverProperties, factory);
-			if (this.maxThreads > 0) {
-				customizeMaxThreads(factory);
-			}
-			if (this.minSpareThreads > 0) {
-				customizeMinThreads(factory);
-			}
-			int maxHttpHeaderSize = (serverProperties.getMaxHttpHeaderSize() > 0
-					? serverProperties.getMaxHttpHeaderSize() : this.maxHttpHeaderSize);
-			if (maxHttpHeaderSize > 0) {
-				customizeMaxHttpHeaderSize(factory, maxHttpHeaderSize);
-			}
-			if (this.maxHttpPostSize > 0) {
-				customizeMaxHttpPostSize(factory, this.maxHttpPostSize);
-			}
-			if (this.accesslog.enabled) {
-				customizeAccessLog(factory);
-			}
-			if (getUriEncoding() != null) {
-				factory.setUriEncoding(getUriEncoding());
-			}
-			if (serverProperties.getConnectionTimeout() != null) {
-				customizeConnectionTimeout(factory,
-						serverProperties.getConnectionTimeout());
-			}
-			if (this.redirectContextRoot != null) {
-				customizeRedirectContextRoot(factory, this.redirectContextRoot);
-			}
-			if (this.maxConnections > 0) {
-				customizeMaxConnections(factory);
-			}
-			if (this.acceptCount > 0) {
-				customizeAcceptCount(factory);
-			}
-			if (!ObjectUtils.isEmpty(this.additionalTldSkipPatterns)) {
-				factory.getTldSkipPatterns().addAll(this.additionalTldSkipPatterns);
-			}
-		}
-
-		private void customizeAcceptCount(TomcatEmbeddedServletContainerFactory factory) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-
-				@Override
-				public void customize(Connector connector) {
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractProtocol) {
-						AbstractProtocol<?> protocol = (AbstractProtocol<?>) handler;
-						protocol.setBacklog(Tomcat.this.acceptCount);
-					}
-				}
-
-			});
-		}
-
-		private void customizeMaxConnections(
-				TomcatEmbeddedServletContainerFactory factory) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-
-				@Override
-				public void customize(Connector connector) {
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractProtocol) {
-						AbstractProtocol<?> protocol = (AbstractProtocol<?>) handler;
-						protocol.setMaxConnections(Tomcat.this.maxConnections);
-					}
-				}
-
-			});
-		}
-
-		private void customizeConnectionTimeout(
-				TomcatEmbeddedServletContainerFactory factory,
-				final int connectionTimeout) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-
-				@Override
-				public void customize(Connector connector) {
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractProtocol) {
-						AbstractProtocol<?> protocol = (AbstractProtocol<?>) handler;
-						protocol.setConnectionTimeout(connectionTimeout);
-					}
-				}
-
-			});
-		}
-
-		private void customizeRemoteIpValve(ServerProperties properties,
-				TomcatEmbeddedServletContainerFactory factory) {
-			String protocolHeader = getProtocolHeader();
-			String remoteIpHeader = getRemoteIpHeader();
-			// For back compatibility the valve is also enabled if protocol-header is set
-			if (StringUtils.hasText(protocolHeader) || StringUtils.hasText(remoteIpHeader)
-					|| properties.getOrDeduceUseForwardHeaders()) {
-				RemoteIpValve valve = new RemoteIpValve();
-				valve.setProtocolHeader(StringUtils.hasLength(protocolHeader)
-						? protocolHeader : "X-Forwarded-Proto");
-				if (StringUtils.hasLength(remoteIpHeader)) {
-					valve.setRemoteIpHeader(remoteIpHeader);
-				}
-				// The internal proxies default to a white list of "safe" internal IP
-				// addresses
-				valve.setInternalProxies(getInternalProxies());
-				valve.setPortHeader(getPortHeader());
-				valve.setProtocolHeaderHttpsValue(getProtocolHeaderHttpsValue());
-				// ... so it's safe to add this valve by default.
-				factory.addEngineValves(valve);
-			}
-		}
-
-		@SuppressWarnings("rawtypes")
-		private void customizeMaxThreads(TomcatEmbeddedServletContainerFactory factory) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-				@Override
-				public void customize(Connector connector) {
-
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractProtocol) {
-						AbstractProtocol protocol = (AbstractProtocol) handler;
-						protocol.setMaxThreads(Tomcat.this.maxThreads);
-					}
-
-				}
-			});
-		}
-
-		@SuppressWarnings("rawtypes")
-		private void customizeMinThreads(TomcatEmbeddedServletContainerFactory factory) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-				@Override
-				public void customize(Connector connector) {
-
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractProtocol) {
-						AbstractProtocol protocol = (AbstractProtocol) handler;
-						protocol.setMinSpareThreads(Tomcat.this.minSpareThreads);
-					}
-
-				}
-			});
-		}
-
-		@SuppressWarnings("rawtypes")
-		private void customizeMaxHttpHeaderSize(
-				TomcatEmbeddedServletContainerFactory factory,
-				final int maxHttpHeaderSize) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-
-				@Override
-				public void customize(Connector connector) {
-					ProtocolHandler handler = connector.getProtocolHandler();
-					if (handler instanceof AbstractHttp11Protocol) {
-						AbstractHttp11Protocol protocol = (AbstractHttp11Protocol) handler;
-						protocol.setMaxHttpHeaderSize(maxHttpHeaderSize);
-					}
-				}
-
-			});
-		}
-
-		private void customizeMaxHttpPostSize(
-				TomcatEmbeddedServletContainerFactory factory,
-				final int maxHttpPostSize) {
-			factory.addConnectorCustomizers(new TomcatConnectorCustomizer() {
-
-				@Override
-				public void customize(Connector connector) {
-					connector.setMaxPostSize(maxHttpPostSize);
-				}
-
-			});
-		}
-
-		private void customizeAccessLog(TomcatEmbeddedServletContainerFactory factory) {
-			AccessLogValve valve = new AccessLogValve();
-			valve.setPattern(this.accesslog.getPattern());
-			valve.setDirectory(this.accesslog.getDirectory());
-			valve.setPrefix(this.accesslog.getPrefix());
-			valve.setSuffix(this.accesslog.getSuffix());
-			valve.setRenameOnRotate(this.accesslog.isRenameOnRotate());
-			valve.setRequestAttributesEnabled(
-					this.accesslog.isRequestAttributesEnabled());
-			valve.setRotatable(this.accesslog.isRotate());
-			valve.setBuffered(this.accesslog.isBuffered());
-			factory.addEngineValves(valve);
-		}
-
-		private void customizeRedirectContextRoot(
-				TomcatEmbeddedServletContainerFactory factory,
-				final boolean redirectContextRoot) {
-			factory.addContextCustomizers(new TomcatContextCustomizer() {
-
-				@Override
-				public void customize(Context context) {
-					context.setMapperContextRootRedirectEnabled(redirectContextRoot);
-				}
-
-			});
 		}
 
 		public static class Accesslog {
@@ -1072,126 +760,6 @@ public class ServerProperties
 			this.selectors = selectors;
 		}
 
-		void customizeJetty(final ServerProperties serverProperties,
-				JettyEmbeddedServletContainerFactory factory) {
-			factory.setUseForwardHeaders(serverProperties.getOrDeduceUseForwardHeaders());
-			if (this.acceptors != null) {
-				factory.setAcceptors(this.acceptors);
-			}
-			if (this.selectors != null) {
-				factory.setSelectors(this.selectors);
-			}
-			if (serverProperties.getMaxHttpHeaderSize() > 0) {
-				customizeMaxHttpHeaderSize(factory,
-						serverProperties.getMaxHttpHeaderSize());
-			}
-			if (this.maxHttpPostSize > 0) {
-				customizeMaxHttpPostSize(factory, this.maxHttpPostSize);
-			}
-
-			if (serverProperties.getConnectionTimeout() != null) {
-				customizeConnectionTimeout(factory,
-						serverProperties.getConnectionTimeout());
-			}
-		}
-
-		private void customizeConnectionTimeout(
-				JettyEmbeddedServletContainerFactory factory,
-				final int connectionTimeout) {
-			factory.addServerCustomizers(new JettyServerCustomizer() {
-
-				@Override
-				public void customize(Server server) {
-					for (org.eclipse.jetty.server.Connector connector : server
-							.getConnectors()) {
-						if (connector instanceof AbstractConnector) {
-							((AbstractConnector) connector)
-									.setIdleTimeout(connectionTimeout);
-						}
-					}
-				}
-
-			});
-		}
-
-		private void customizeMaxHttpHeaderSize(
-				JettyEmbeddedServletContainerFactory factory,
-				final int maxHttpHeaderSize) {
-			factory.addServerCustomizers(new JettyServerCustomizer() {
-
-				@Override
-				public void customize(Server server) {
-					for (org.eclipse.jetty.server.Connector connector : server
-							.getConnectors()) {
-						try {
-							for (ConnectionFactory connectionFactory : connector
-									.getConnectionFactories()) {
-								if (connectionFactory instanceof HttpConfiguration.ConnectionFactory) {
-									customize(
-											(HttpConfiguration.ConnectionFactory) connectionFactory);
-								}
-							}
-						}
-						catch (NoSuchMethodError ex) {
-							customizeOnJetty8(connector, maxHttpHeaderSize);
-						}
-					}
-
-				}
-
-				private void customize(HttpConfiguration.ConnectionFactory factory) {
-					HttpConfiguration configuration = factory.getHttpConfiguration();
-					configuration.setRequestHeaderSize(maxHttpHeaderSize);
-					configuration.setResponseHeaderSize(maxHttpHeaderSize);
-				}
-
-				private void customizeOnJetty8(
-						org.eclipse.jetty.server.Connector connector,
-						int maxHttpHeaderSize) {
-					try {
-						connector.getClass().getMethod("setRequestHeaderSize", int.class)
-								.invoke(connector, maxHttpHeaderSize);
-						connector.getClass().getMethod("setResponseHeaderSize", int.class)
-								.invoke(connector, maxHttpHeaderSize);
-					}
-					catch (Exception ex) {
-						throw new RuntimeException(ex);
-					}
-				}
-
-			});
-		}
-
-		private void customizeMaxHttpPostSize(
-				JettyEmbeddedServletContainerFactory factory, final int maxHttpPostSize) {
-			factory.addServerCustomizers(new JettyServerCustomizer() {
-
-				@Override
-				public void customize(Server server) {
-					setHandlerMaxHttpPostSize(maxHttpPostSize, server.getHandlers());
-				}
-
-				private void setHandlerMaxHttpPostSize(int maxHttpPostSize,
-						Handler... handlers) {
-					for (Handler handler : handlers) {
-						if (handler instanceof ContextHandler) {
-							((ContextHandler) handler)
-									.setMaxFormContentSize(maxHttpPostSize);
-						}
-						else if (handler instanceof HandlerWrapper) {
-							setHandlerMaxHttpPostSize(maxHttpPostSize,
-									((HandlerWrapper) handler).getHandler());
-						}
-						else if (handler instanceof HandlerCollection) {
-							setHandlerMaxHttpPostSize(maxHttpPostSize,
-									((HandlerCollection) handler).getHandlers());
-						}
-					}
-				}
-
-			});
-		}
-
 	}
 
 	public static class Undertow {
@@ -1282,82 +850,6 @@ public class ServerProperties
 			return this.accesslog;
 		}
 
-		void customizeUndertow(final ServerProperties serverProperties,
-				UndertowEmbeddedServletContainerFactory factory) {
-			if (this.bufferSize != null) {
-				factory.setBufferSize(this.bufferSize);
-			}
-			if (this.ioThreads != null) {
-				factory.setIoThreads(this.ioThreads);
-			}
-			if (this.workerThreads != null) {
-				factory.setWorkerThreads(this.workerThreads);
-			}
-			if (this.directBuffers != null) {
-				factory.setDirectBuffers(this.directBuffers);
-			}
-			if (this.accesslog.enabled != null) {
-				factory.setAccessLogEnabled(this.accesslog.enabled);
-			}
-			factory.setAccessLogDirectory(this.accesslog.dir);
-			factory.setAccessLogPattern(this.accesslog.pattern);
-			factory.setAccessLogPrefix(this.accesslog.prefix);
-			factory.setAccessLogSuffix(this.accesslog.suffix);
-			factory.setAccessLogRotate(this.accesslog.rotate);
-			factory.setUseForwardHeaders(serverProperties.getOrDeduceUseForwardHeaders());
-			if (serverProperties.getMaxHttpHeaderSize() > 0) {
-				customizeMaxHttpHeaderSize(factory,
-						serverProperties.getMaxHttpHeaderSize());
-			}
-			if (this.maxHttpPostSize > 0) {
-				customizeMaxHttpPostSize(factory, this.maxHttpPostSize);
-			}
-
-			if (serverProperties.getConnectionTimeout() != null) {
-				customizeConnectionTimeout(factory,
-						serverProperties.getConnectionTimeout());
-			}
-		}
-
-		private void customizeConnectionTimeout(
-				UndertowEmbeddedServletContainerFactory factory,
-				final int connectionTimeout) {
-			factory.addBuilderCustomizers(new UndertowBuilderCustomizer() {
-				@Override
-				public void customize(Builder builder) {
-					builder.setSocketOption(UndertowOptions.NO_REQUEST_TIMEOUT,
-							connectionTimeout);
-				}
-			});
-		}
-
-		private void customizeMaxHttpHeaderSize(
-				UndertowEmbeddedServletContainerFactory factory,
-				final int maxHttpHeaderSize) {
-			factory.addBuilderCustomizers(new UndertowBuilderCustomizer() {
-
-				@Override
-				public void customize(Builder builder) {
-					builder.setServerOption(UndertowOptions.MAX_HEADER_SIZE,
-							maxHttpHeaderSize);
-				}
-
-			});
-		}
-
-		private void customizeMaxHttpPostSize(
-				UndertowEmbeddedServletContainerFactory factory,
-				final long maxHttpPostSize) {
-			factory.addBuilderCustomizers(new UndertowBuilderCustomizer() {
-
-				@Override
-				public void customize(Builder builder) {
-					builder.setServerOption(UndertowOptions.MAX_ENTITY_SIZE,
-							maxHttpPostSize);
-				}
-
-			});
-		}
 
 		public static class Accesslog {
 
@@ -1439,54 +931,6 @@ public class ServerProperties
 				this.rotate = rotate;
 			}
 
-		}
-
-	}
-
-	/**
-	 * {@link ServletContextInitializer} to apply appropriate parts of the {@link Session}
-	 * configuration.
-	 */
-	private static class SessionConfiguringInitializer
-			implements ServletContextInitializer {
-
-		private final Session session;
-
-		SessionConfiguringInitializer(Session session) {
-			this.session = session;
-		}
-
-		@Override
-		public void onStartup(ServletContext servletContext) throws ServletException {
-			if (this.session.getTrackingModes() != null) {
-				servletContext.setSessionTrackingModes(this.session.getTrackingModes());
-			}
-			configureSessionCookie(servletContext.getSessionCookieConfig());
-		}
-
-		private void configureSessionCookie(SessionCookieConfig config) {
-			Cookie cookie = this.session.getCookie();
-			if (cookie.getName() != null) {
-				config.setName(cookie.getName());
-			}
-			if (cookie.getDomain() != null) {
-				config.setDomain(cookie.getDomain());
-			}
-			if (cookie.getPath() != null) {
-				config.setPath(cookie.getPath());
-			}
-			if (cookie.getComment() != null) {
-				config.setComment(cookie.getComment());
-			}
-			if (cookie.getHttpOnly() != null) {
-				config.setHttpOnly(cookie.getHttpOnly());
-			}
-			if (cookie.getSecure() != null) {
-				config.setSecure(cookie.getSecure());
-			}
-			if (cookie.getMaxAge() != null) {
-				config.setMaxAge(cookie.getMaxAge());
-			}
 		}
 
 	}
